+++ bitloops_cli/src/engine/mod.rs
@@ -3,9 +3,11 @@ pub mod db_status;
 pub mod git_operations;
 pub mod history;
 pub mod hooks;
+pub mod infra;
 pub mod lifecycle;
 pub mod logging;
 pub mod paths;
+pub mod providers;
 pub mod semantic_features;
 pub mod session;
 pub mod settings;
